refactor: extract FieldMapperService from SchemaCore
Move apply_field_mappers (and its circular-redirect source-resolution
logic) out of SchemaCore into a dedicated FieldMapperService in
src/schema/field_mapper.rs. SchemaCore now holds a FieldMapperService
and delegates during set_schema_state when a schema is approved.

Pure refactor — no behavior change. Drops core.rs by 84 LOC and
documents the schema-expansion workflow and circular-redirect rationale
in the new module.
