+++ src/schema/mod.rs
@@ -23,12 +23,14 @@
 
 // Internal modules
 pub mod core;
+pub mod field_mapper;
 pub mod persistence;
 pub mod schema_types;
 pub mod types;
 
 // Public re-exports
 pub use core::SchemaCore;
+pub use field_mapper::FieldMapperService;
 pub use schema_types::{SchemaState, SchemaWithState};
 pub use types::{Schema, SchemaError};
